Run Flask alongside messagebroker to be able to create an export config endpoint
