+++ docker-compose.yml
@@ -13,6 +13,15 @@ services:
       OIDC_TOKEN_ENDPOINT: ${OIDC_TOKEN_ENDPOINT}
       OIDC_CLIENT_ID: ${OIDC_CLIENT_ID}
       OIDC_CLIENT_SECRET: ${OIDC_CLIENT_SECRET}
+
+      UWSGI_HTTP: ":8000"
+      UWSGI_MODULE: "gobexport.wsgi"
+      UWSGI_CALLABLE: "app"
+      UWSGI_MASTER: 1
+      UWSGI_PROCESSES: "4"
+      UWSGI_LAZY: "1"
+      UWSGI_ENABLE_THREADS: ""
+      UWSGI_THREADS: ""
     env_file: .env
 
 networks:
