+++ src/gobexport/wsgi.py
@@ -1,0 +1,4 @@
+from gobexport.app import get_app
+
+# Run the app with uWSGI
+app = get_app()
