+++ src/requirements.txt
@@ -6,9 +6,10 @@ chardet==3.0.4
 Click==7.0
 coverage==4.5.1
 datapunt-config-loader==1.1.0
--e git+https://github.com/Amsterdam/objectstore.git@v1.0.1#egg=objectstore
 debtcollector==1.20.0
 flake8==3.5.0
+Flask==1.1.2
+Flask-Cors==3.0.8
 freezegun==0.3.15
 GDAL==2.4.4
 GeoAlchemy2==0.5.0
@@ -45,11 +46,9 @@ python-keystoneclient==3.18.0
 python-swiftclient==3.6.0
 pytz==2018.7
 PyYAML==4.2b1
-requests==2.22.0
 rfc3986==1.1.0
 Shapely==1.6.4.post2
 six==1.11.0
 SQLAlchemy==1.3.3
 stevedore==1.30.0
-urllib3==1.25.7
 wrapt==1.10.11
